Add getindex for Chain & Square lattice
getindex will return a serial number (column-wise for square)
this is useful while trying to gather the order of
a set of operators on lattice.

Add utils: fastmod1, this is a generated function evals mod1 in
compile time, 40x faster than built mod1. And `LessEqual` number
a periodic integer with upper bound, this should accelerate some
operation as well, since we will know the prior info of the bound.